ENT-942 Implement an abstraction layer for SoftwareSecurePhotoVerification class methods
