+++ common/djangoapps/student/views/dashboard.py
@@ -11,10 +11,10 @@ from completion.utilities import get_key_to_last_completed_course_block
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.core.urlresolvers import NoReverseMatch, reverse, reverse_lazy
+from django.core.urlresolvers import reverse
 from django.shortcuts import redirect
 from django.utils.translation import ugettext as _
-from django.views.decorators.csrf import csrf_exempt, ensure_csrf_cookie
+from django.views.decorators.csrf import ensure_csrf_cookie
 
 from opaque_keys.edx.keys import CourseKey
 from pytz import UTC
@@ -27,7 +27,7 @@ from courseware.access import has_access
 from edxmako.shortcuts import render_to_response, render_to_string
 from entitlements.models import CourseEntitlement
 from lms.djangoapps.commerce.utils import EcommerceService  # pylint: disable=import-error
-from lms.djangoapps.verify_student.models import SoftwareSecurePhotoVerification  # pylint: disable=import-error
+from lms.djangoapps.verify_student.services import IDVerificationService
 from openedx.core.djangoapps import monitoring_utils
 from openedx.core.djangoapps.catalog.utils import (
     get_programs,
@@ -339,6 +339,9 @@ def is_course_blocked(request, redeemed_registration_codes, course_key):
 
 
 def get_verification_error_reasons_for_display(verification_error_codes):
+    """
+    Returns the display text for the given verification error codes.
+    """
     verification_errors = []
     verification_error_map = {
         'photos_mismatched': _('Photos are mismatched'),
@@ -713,7 +716,7 @@ def student_dashboard(request):
 
     # Verification Attempts
     # Used to generate the "you must reverify for course x" banner
-    verification_status, verification_error_codes = SoftwareSecurePhotoVerification.user_status(user)
+    verification_status, verification_error_codes = IDVerificationService.user_status(user)
     verification_errors = get_verification_error_reasons_for_display(verification_error_codes)
 
     # Gets data for midcourse reverifications, if any are necessary or have failed
