+++ lms/djangoapps/instructor/tests/test_certificates.py
@@ -32,7 +32,7 @@ from lms.djangoapps.certificates.tests.factories import (
 from course_modes.models import CourseMode
 from courseware.tests.factories import GlobalStaffFactory, InstructorFactory, UserFactory
 from lms.djangoapps.grades.tests.utils import mock_passing_grade
-from lms.djangoapps.verify_student.models import SoftwareSecurePhotoVerification
+from lms.djangoapps.verify_student.services import IDVerificationService
 from lms.djangoapps.verify_student.tests.factories import SoftwareSecurePhotoVerificationFactory
 from student.models import CourseEnrollment
 from xmodule.modulestore.tests.django_utils import SharedModuleStoreTestCase
@@ -398,10 +398,9 @@ class CertificatesInstructorApiTest(SharedModuleStoreTestCase):
 
             # Create and assert user's ID verification record.
             SoftwareSecurePhotoVerificationFactory.create(user=self.user, status=id_verification_status)
-            actual_verification_status = SoftwareSecurePhotoVerification.verification_status_for_user(
+            actual_verification_status = IDVerificationService.verification_status_for_user(
                 self.user,
-                self.course.id,
-                enrollment.mode,
+                enrollment.mode
             )
             self.assertEquals(actual_verification_status, verification_output)
