+++ lms/djangoapps/instructor_analytics/tests/test_basic.py
@@ -184,7 +184,7 @@ class TestAnalyticsBasic(ModuleStoreTestCase):
         # is returned by verification and enrollment code
         with patch("student.models.CourseEnrollment.enrollment_mode_for_user") as enrollment_patch:
             with patch(
-                "lms.djangoapps.verify_student.models.SoftwareSecurePhotoVerification.verification_status_for_user"
+                "lms.djangoapps.verify_student.services.IDVerificationService.verification_status_for_user"
             ) as verify_patch:
                 enrollment_patch.return_value = ["verified"]
                 verify_patch.return_value = "dummy verification status"
