+++ lms/djangoapps/verify_student/models.py
@@ -27,13 +27,11 @@ from django.core.urlresolvers import reverse
 from django.db import models
 from django.dispatch import receiver
 from django.utils.functional import cached_property
-from django.utils.translation import ugettext as _
 from django.utils.translation import ugettext_lazy
 from model_utils import Choices
 from model_utils.models import StatusModel, TimeStampedModel
 from opaque_keys.edx.django.models import CourseKeyField
 
-from course_modes.models import CourseMode
 from lms.djangoapps.verify_student.ssencrypt import (
     encrypt_and_encode,
     generate_signed_message,
@@ -41,9 +39,9 @@ from lms.djangoapps.verify_student.ssencrypt import (
     rsa_encrypt
 )
 from openedx.core.djangoapps.signals.signals import LEARNER_NOW_VERIFIED
-from openedx.core.djangoapps.site_configuration import helpers as configuration_helpers
 from openedx.core.storage import get_storage
 
+from .utils import earliest_allowed_verification_date
 
 log = logging.getLogger(__name__)
 
@@ -192,213 +190,6 @@ class PhotoVerification(StatusModel):
         app_label = "verify_student"
         abstract = True
         ordering = ['-created_at']
-
-    ##### Methods listed in the order you'd typically call them
-    @classmethod
-    def _earliest_allowed_date(cls):
-        """
-        Returns the earliest allowed date given the settings
-
-        """
-        days_good_for = settings.VERIFY_STUDENT["DAYS_GOOD_FOR"]
-        return datetime.now(pytz.UTC) - timedelta(days=days_good_for)
-
-    @classmethod
-    def user_is_verified(cls, user, earliest_allowed_date=None):
-        """
-        Return whether or not a user has satisfactorily proved their identity.
-        Depending on the policy, this can expire after some period of time, so
-        a user might have to renew periodically.
-
-        This will check for the user's *initial* verification.
-        """
-        return cls.verified_query(earliest_allowed_date).filter(user=user).exists()
-
-    @classmethod
-    def verified_query(cls, earliest_allowed_date=None):
-        """
-        Return a query set for all records with 'approved' state
-        that are still valid according to the earliest_allowed_date
-        value or policy settings.
-        """
-        return cls.objects.filter(
-            status="approved",
-            created_at__gte=(earliest_allowed_date or cls._earliest_allowed_date()),
-        )
-
-    @classmethod
-    def verification_valid_or_pending(cls, user, earliest_allowed_date=None, queryset=None):
-        """
-        Check whether the user has a complete verification attempt that is
-        or *might* be good. This means that it's approved, been submitted,
-        or would have been submitted but had an non-user error when it was
-        being submitted.
-        It's basically any situation in which the user has signed off on
-        the contents of the attempt, and we have not yet received a denial.
-        This will check for the user's *initial* verification.
-
-        Arguments:
-            user:
-            earliest_allowed_date: earliest allowed date given in the
-                settings
-            queryset: If a queryset is provided, that will be used instead
-                of hitting the database.
-
-        Returns:
-            queryset: queryset of 'PhotoVerification' sorted by 'created_at' in
-            descending order.
-        """
-
-        valid_statuses = ['submitted', 'approved', 'must_retry']
-
-        if queryset is None:
-            queryset = cls.objects.filter(user=user)
-
-        return queryset.filter(
-            status__in=valid_statuses,
-            created_at__gte=(
-                earliest_allowed_date
-                or cls._earliest_allowed_date()
-            )
-        ).order_by('-created_at')
-
-    @classmethod
-    def get_expiration_datetime(cls, user, queryset=None):
-        """
-        Check whether the user has an approved verification and return the
-        "expiration_datetime" of most recent "approved" verification.
-
-        Arguments:
-            user (Object): User
-            queryset: If a queryset is provided, that will be used instead
-                of hitting the database.
-
-        Returns:
-            expiration_datetime: expiration_datetime of most recent "approved"
-            verification.
-        """
-        if queryset is None:
-            queryset = cls.objects.filter(user=user)
-
-        photo_verification = queryset.filter(status='approved').first()
-        if photo_verification:
-            return photo_verification.expiration_datetime
-
-    @classmethod
-    def user_has_valid_or_pending(cls, user, earliest_allowed_date=None, queryset=None):
-        """
-        Check whether the user has an active or pending verification attempt
-
-        Returns:
-            bool: True or False according to existence of valid verifications
-        """
-        return cls.verification_valid_or_pending(user, earliest_allowed_date, queryset).exists()
-
-    @classmethod
-    def active_for_user(cls, user):
-        """
-        Return the most recent PhotoVerification that is marked ready (i.e. the
-        user has said they're set, but we haven't submitted anything yet).
-
-        This checks for the original verification.
-        """
-        # This should only be one at the most, but just in case we create more
-        # by mistake, we'll grab the most recently created one.
-        active_attempts = cls.objects.filter(user=user, status='ready').order_by('-created_at')
-        if active_attempts:
-            return active_attempts[0]
-        else:
-            return None
-
-    @classmethod
-    def user_status(cls, user):
-        """
-        Returns the status of the user based on their past verification attempts
-
-        If no such verification exists, returns 'none'
-        If verification has expired, returns 'expired'
-        If the verification has been approved, returns 'approved'
-        If the verification process is still ongoing, returns 'pending'
-        If the verification has been denied and the user must resubmit photos, returns 'must_reverify'
-
-        This checks initial verifications
-        """
-        status = 'none'
-        error_msg = ''
-
-        if cls.user_is_verified(user):
-            status = 'approved'
-
-        elif cls.user_has_valid_or_pending(user):
-            # user_has_valid_or_pending does include 'approved', but if we are
-            # here, we know that the attempt is still pending
-            status = 'pending'
-
-        else:
-            # we need to check the most recent attempt to see if we need to ask them to do
-            # a retry
-            try:
-                attempts = cls.objects.filter(user=user).order_by('-updated_at')
-                attempt = attempts[0]
-            except IndexError:
-                # we return 'none'
-
-                return ('none', error_msg)
-
-            if attempt.created_at < cls._earliest_allowed_date():
-                return (
-                    'expired',
-                    _("Your {platform_name} verification has expired.").format(
-                        platform_name=configuration_helpers.get_value('platform_name', settings.PLATFORM_NAME),
-                    )
-                )
-
-            # If someone is denied their original verification attempt, they can try to reverify.
-            if attempt.status == 'denied':
-                status = 'must_reverify'
-
-            if attempt.error_msg:
-                error_msg = attempt.parsed_error_msg()
-
-        return (status, error_msg)
-
-    @classmethod
-    def verification_for_datetime(cls, deadline, candidates):
-        """Find a verification in a set that applied during a particular datetime.
-
-        A verification is considered "active" during a datetime if:
-        1) The verification was created before the datetime, and
-        2) The verification is set to expire after the datetime.
-
-        Note that verification status is *not* considered here,
-        just the start/expire dates.
-
-        If multiple verifications were active at the deadline,
-        returns the most recently created one.
-
-        Arguments:
-            deadline (datetime): The datetime at which the verification applied.
-                If `None`, then return the most recently created candidate.
-            candidates (list of `PhotoVerification`s): Potential verifications to search through.
-
-        Returns:
-            PhotoVerification: A photo verification that was active at the deadline.
-                If no verification was active, return None.
-
-        """
-        if len(candidates) == 0:
-            return None
-
-        # If there's no deadline, then return the most recently created verification
-        if deadline is None:
-            return candidates[0]
-
-        # Otherwise, look for a verification that was in effect at the deadline,
-        # preferring recent verifications.
-        # If no such verification is found, implicitly return `None`
-        for verification in candidates:
-            if verification.active_at_datetime(deadline):
-                return verification
 
     @property
     def expiration_datetime(self):
@@ -642,7 +433,7 @@ class SoftwareSecurePhotoVerification(PhotoVerification):
             user=user,
             status__in=["submitted", "approved"],
             created_at__gte=(
-                earliest_allowed_date or cls._earliest_allowed_date()
+                earliest_allowed_date or earliest_allowed_verification_date()
             )
         ).exclude(photo_id_key='')
 
@@ -973,34 +764,6 @@ class SoftwareSecurePhotoVerification(PhotoVerification):
         log.debug("Return message:\n\n{}\n\n".format(response.text))
 
         return response
-
-    @classmethod
-    def verification_status_for_user(cls, user, course_id, user_enrollment_mode, user_is_verified=None):
-        """
-        Returns the verification status for use in grade report.
-        """
-        if user_enrollment_mode not in CourseMode.VERIFIED_MODES:
-            return 'N/A'
-
-        if user_is_verified is None:
-            user_is_verified = cls.user_is_verified(user)
-
-        if not user_is_verified:
-            return 'Not ID Verified'
-        else:
-            return 'ID Verified'
-
-    @classmethod
-    def is_verification_expiring_soon(cls, expiration_datetime):
-        """
-        Returns True if verification is expiring within EXPIRING_SOON_WINDOW.
-        """
-        if expiration_datetime:
-            if (expiration_datetime - datetime.now(pytz.UTC)).days <= settings.VERIFY_STUDENT.get(
-                    "EXPIRING_SOON_WINDOW"):
-                return True
-
-        return False
 
 
 class VerificationDeadline(TimeStampedModel):
