+++ lms/djangoapps/verify_student/tests/test_models.py
@@ -12,8 +12,6 @@ from freezegun import freeze_time
 from mock import patch
 from nose.tools import (  # pylint: disable=no-name-in-module
     assert_equals,
-    assert_false,
-    assert_is_none,
     assert_raises,
     assert_true
 )
@@ -29,7 +27,6 @@ from lms.djangoapps.verify_student.models import (
 from openedx.core.djangolib.testing.utils import CacheIsolationTestCase
 from student.tests.factories import UserFactory
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
-from xmodule.modulestore.tests.factories import CourseFactory
 
 FAKE_SETTINGS = {
     "SOFTWARE_SECURE": {
@@ -233,116 +230,6 @@ class TestPhotoVerification(MockS3Mixin, ModuleStoreTestCase):
 
         self.assertEqual(attempt.photo_id_key, "fake-photo-id-key")
 
-    def test_active_for_user(self):
-        """
-        Make sure we can retrive a user's active (in progress) verification
-        attempt.
-        """
-        user = UserFactory.create()
-
-        # This user has no active at the moment...
-        assert_is_none(SoftwareSecurePhotoVerification.active_for_user(user))
-
-        # Create an attempt and mark it ready...
-        attempt = SoftwareSecurePhotoVerification(user=user)
-        attempt.mark_ready()
-        assert_equals(attempt, SoftwareSecurePhotoVerification.active_for_user(user))
-
-        # A new user won't see this...
-        user2 = UserFactory.create()
-        user2.save()
-        assert_is_none(SoftwareSecurePhotoVerification.active_for_user(user2))
-
-        # If it's got a different status, it doesn't count
-        for status in ["submitted", "must_retry", "approved", "denied"]:
-            attempt.status = status
-            attempt.save()
-            assert_is_none(SoftwareSecurePhotoVerification.active_for_user(user))
-
-        # But if we create yet another one and mark it ready, it passes again.
-        attempt_2 = SoftwareSecurePhotoVerification(user=user)
-        attempt_2.mark_ready()
-        assert_equals(attempt_2, SoftwareSecurePhotoVerification.active_for_user(user))
-
-        # And if we add yet another one with a later created time, we get that
-        # one instead. We always want the most recent attempt marked ready()
-        attempt_3 = SoftwareSecurePhotoVerification(
-            user=user,
-            created_at=attempt_2.created_at + timedelta(days=1)
-        )
-        attempt_3.save()
-
-        # We haven't marked attempt_3 ready yet, so attempt_2 still wins
-        assert_equals(attempt_2, SoftwareSecurePhotoVerification.active_for_user(user))
-
-        # Now we mark attempt_3 ready and expect it to come back
-        attempt_3.mark_ready()
-        assert_equals(attempt_3, SoftwareSecurePhotoVerification.active_for_user(user))
-
-    def test_user_is_verified(self):
-        """
-        Test to make sure we correctly answer whether a user has been verified.
-        """
-        user = UserFactory.create()
-        attempt = SoftwareSecurePhotoVerification(user=user)
-        attempt.save()
-
-        # If it's any of these, they're not verified...
-        for status in ["created", "ready", "denied", "submitted", "must_retry"]:
-            attempt.status = status
-            attempt.save()
-            assert_false(SoftwareSecurePhotoVerification.user_is_verified(user), status)
-
-        attempt.status = "approved"
-        attempt.save()
-        assert_true(SoftwareSecurePhotoVerification.user_is_verified(user), attempt.status)
-
-    def test_user_has_valid_or_pending(self):
-        """
-        Determine whether we have to prompt this user to verify, or if they've
-        already at least initiated a verification submission.
-        """
-        user = UserFactory.create()
-        attempt = SoftwareSecurePhotoVerification(user=user)
-
-        # If it's any of these statuses, they don't have anything outstanding
-        for status in ["created", "ready", "denied"]:
-            attempt.status = status
-            attempt.save()
-            assert_false(SoftwareSecurePhotoVerification.user_has_valid_or_pending(user), status)
-
-        # Any of these, and we are. Note the benefit of the doubt we're giving
-        # -- must_retry, and submitted both count until we hear otherwise
-        for status in ["submitted", "must_retry", "approved"]:
-            attempt.status = status
-            attempt.save()
-            assert_true(SoftwareSecurePhotoVerification.user_has_valid_or_pending(user), status)
-
-    def test_user_status(self):
-        # test for correct status when no error returned
-        user = UserFactory.create()
-        status = SoftwareSecurePhotoVerification.user_status(user)
-        self.assertEquals(status, ('none', ''))
-
-        # test for when one has been created
-        attempt = SoftwareSecurePhotoVerification.objects.create(user=user, status='approved')
-        status = SoftwareSecurePhotoVerification.user_status(user)
-        self.assertEquals(status, ('approved', ''))
-
-        # create another one for the same user, make sure the right one is
-        # returned
-        SoftwareSecurePhotoVerification.objects.create(
-            user=user, status='denied', error_msg='[{"photoIdReasons": ["Not provided"]}]'
-        )
-        status = SoftwareSecurePhotoVerification.user_status(user)
-        self.assertEquals(status, ('approved', ''))
-
-        # now delete the first one and verify that the denial is being handled
-        # properly
-        attempt.delete()
-        status = SoftwareSecurePhotoVerification.user_status(user)
-        self.assertEquals(status, ('must_reverify', ['id_image_missing']))
-
     # pylint: disable=line-too-long
     def test_parse_error_msg_success(self):
         user = UserFactory.create()
@@ -383,87 +270,6 @@ class TestPhotoVerification(MockS3Mixin, ModuleStoreTestCase):
         attempt.created_at = attempt.created_at - timedelta(days=settings.VERIFY_STUDENT["DAYS_GOOD_FOR"])
         attempt.save()
         self.assertFalse(attempt.active_at_datetime(datetime.now(pytz.UTC) + timedelta(days=1)))
-
-    def test_verification_for_datetime(self):
-        user = UserFactory.create()
-        now = datetime.now(pytz.UTC)
-
-        # No attempts in the query set, so should return None
-        query = SoftwareSecurePhotoVerification.objects.filter(user=user)
-        result = SoftwareSecurePhotoVerification.verification_for_datetime(now, query)
-        self.assertIs(result, None)
-
-        # Should also return None if no deadline specified
-        query = SoftwareSecurePhotoVerification.objects.filter(user=user)
-        result = SoftwareSecurePhotoVerification.verification_for_datetime(None, query)
-        self.assertIs(result, None)
-
-        # Make an attempt
-        attempt = SoftwareSecurePhotoVerification.objects.create(user=user)
-
-        # Before the created date, should get no results
-        before = attempt.created_at - timedelta(seconds=1)
-        query = SoftwareSecurePhotoVerification.objects.filter(user=user)
-        result = SoftwareSecurePhotoVerification.verification_for_datetime(before, query)
-        self.assertIs(result, None)
-
-        # Immediately after the created date, should get the attempt
-        after_created = attempt.created_at + timedelta(seconds=1)
-        query = SoftwareSecurePhotoVerification.objects.filter(user=user)
-        result = SoftwareSecurePhotoVerification.verification_for_datetime(after_created, query)
-        self.assertEqual(result, attempt)
-
-        # If no deadline specified, should return first available
-        query = SoftwareSecurePhotoVerification.objects.filter(user=user)
-        result = SoftwareSecurePhotoVerification.verification_for_datetime(None, query)
-        self.assertEqual(result, attempt)
-
-        # Immediately before the expiration date, should get the attempt
-        expiration = attempt.created_at + timedelta(days=settings.VERIFY_STUDENT["DAYS_GOOD_FOR"])
-        before_expiration = expiration - timedelta(seconds=1)
-        query = SoftwareSecurePhotoVerification.objects.filter(user=user)
-        result = SoftwareSecurePhotoVerification.verification_for_datetime(before_expiration, query)
-        self.assertEqual(result, attempt)
-
-        # Immediately after the expiration date, should not get the attempt
-        attempt.created_at = attempt.created_at - timedelta(days=settings.VERIFY_STUDENT["DAYS_GOOD_FOR"])
-        attempt.save()
-        after = datetime.now(pytz.UTC) + timedelta(days=1)
-        query = SoftwareSecurePhotoVerification.objects.filter(user=user)
-        result = SoftwareSecurePhotoVerification.verification_for_datetime(after, query)
-        self.assertIs(result, None)
-
-        # Create a second attempt in the same window
-        second_attempt = SoftwareSecurePhotoVerification.objects.create(user=user)
-
-        # Now we should get the newer attempt
-        deadline = second_attempt.created_at + timedelta(days=1)
-        query = SoftwareSecurePhotoVerification.objects.filter(user=user)
-        result = SoftwareSecurePhotoVerification.verification_for_datetime(deadline, query)
-        self.assertEqual(result, second_attempt)
-
-    @ddt.unpack
-    @ddt.data(
-        {'enrollment_mode': 'honor', 'status': None, 'output': 'N/A'},
-        {'enrollment_mode': 'audit', 'status': None, 'output': 'N/A'},
-        {'enrollment_mode': 'verified', 'status': False, 'output': 'Not ID Verified'},
-        {'enrollment_mode': 'verified', 'status': True, 'output': 'ID Verified'},
-    )
-    def test_verification_status_for_user(self, enrollment_mode, status, output):
-        """
-        Verify verification_status_for_user returns correct status.
-        """
-        user = UserFactory.create()
-        course = CourseFactory.create()
-
-        with patch(
-            'lms.djangoapps.verify_student.models.SoftwareSecurePhotoVerification.user_is_verified'
-        ) as mock_verification:
-
-            mock_verification.return_value = status
-
-            status = SoftwareSecurePhotoVerification.verification_status_for_user(user, course.id, enrollment_mode)
-            self.assertEqual(status, output)
 
     def test_initial_verification_for_user(self):
         """Test that method 'get_initial_verification' of model
